Move ByokFactory to common
